# backend/biosim_server/api/main.py
import logging
import os
import uuid
from contextlib import asynccontextmanager
from datetime import datetime, UTC, timedelta
from typing import AsyncGenerator, Optional

import dotenv
import uvicorn
from fastapi import FastAPI, File, UploadFile, Query, APIRouter, Depends, HTTPException, Response
from fastapi.openapi.docs import get_swagger_ui_html
from starlette.middleware.cors import CORSMiddleware
from starlette.responses import HTMLResponse

from biosim_server.biosim_omex import OmexFile, get_cached_omex_file_from_upload
from biosim_server.biosim_runs import BiosimulatorVersion
from biosim_server.biosim_verify import CompareSettings
from biosim_server.compatibility import compatibility_router
from biosim_server.simulations import run_summary_router, simulations_router
from biosim_server.projects.router import router as projects_router
from biosim_server.common.auth import AuthenticatedUser, get_current_user
from biosim_server.common.auth.auth0 import JwksCache, get_jwks_cache
from biosim_server.common.auth.discovery import warm_discovery_cache
from biosim_server.common.auth.roles import require_owner_or_admin
from biosim_server.common.ratelimit import workflow_rate_limit
from biosim_server.rbac_demo.router import router as rbac_demo_router
from biosim_server.users.router import router as users_router
from biosim_server.biosim_verify.models import VerifyWorkflowOutput, VerifyWorkflowStatus
from biosim_server.biosim_verify.omex_verify_workflow import OmexVerifyWorkflow, OmexVerifyWorkflowInput
from biosim_server.biosim_verify.runs_verify_workflow import RunsVerifyWorkflowInput, RunsVerifyWorkflow
from biosim_server.config import get_local_cache_dir, get_settings
from biosim_server.dependencies import get_file_service, get_temporal_client, init_standalone, shutdown_standalone, \
    get_biosim_service, get_omex_database_service, get_mongo_client
from biosim_server.log_config import setup_logging
from biosim_server.version import __version__

logger = logging.getLogger(__name__)
setup_logging(logger)

# -- load dev env -- #
REPO_ROOT = os.path.dirname(os.path.dirname(__file__))
DEV_ENV_PATH = os.path.join(REPO_ROOT, 'assets', 'dev', 'config', '.dev_env')
dotenv.load_dotenv(DEV_ENV_PATH)  # NOTE: create an env config at this filepath if dev

# -- constraints -- #
APP_VERSION = __version__
APP_TITLE = "biosim-server"
# Built-in CORS origins. Two groups, both stable and not deploy-specific:
#  - local-dev loopbacks (127.0.0.1 / localhost on common dev ports)
#  - cross-org trusted services that call this API (biosimulators.*,
#    run.biosimulations.*, etc.)
# Deployment-specific frontend origins (e.g. https://biosim.biosimulations.org,
# https://biosim-dev.biosimulations.org) belong in CORS_EXTRA_ORIGINS instead
# of this list — keeps overlay/deploy URLs explicit and reviewable per-env.
APP_ORIGINS = [
    'http://127.0.0.1:8000',
    'http://127.0.0.1:4200',
    'http://127.0.0.1:4201',
    'http://127.0.0.1:4202',
    'http://localhost:4200',
    'http://localhost:4201',
    'http://localhost:4202',
    'http://localhost:8000',
    'http://localhost:3000',
    'http://localhost:3001',
    'https://biosimulators.org',
    'https://www.biosimulators.org',
    'https://biosimulators.dev',
    'https://www.biosimulators.dev',
    'https://run.biosimulations.dev',
    'https://run.biosimulations.org',
    'https://biosimulations.dev',
    'https://biosimulations.org',
    'https://bio.libretexts.org',
    'https://biochecknet.biosimulations.org'
]

# Deployment-specific origins from env (comma-separated). Empty by default;
# overlay ConfigMaps supply this per cluster/host.
_extra = os.environ.get('CORS_EXTRA_ORIGINS', '').strip()
if _extra:
    APP_ORIGINS = APP_ORIGINS + [o.strip() for o in _extra.split(',') if o.strip()]
APP_SERVERS: list[dict[str, str]] = [
    # {
    #     "url": "https://biochecknet.biosimulations.org",
    #     "description": "Production server"
    # },
    # {
    #     "url": "http://localhost:3001",
    #     "description": "Main Development server"
    # },
    # {
    #     "url": "http://localhost:8000",
    #     "description": "Alternate Development server"
    # }
]

# -- app components -- #

router = APIRouter()


def _validate_auth0_configuration() -> None:
    """
    Startup gate: refuse to serve traffic with an unusable Auth0 configuration.

    Replaces _warn_if_auth0_misconfigured(), which both under-reacted (the pod
    started anyway, reported healthy, and failed every authenticated request)
    and misdescribed the failure -- it promised a 401, which has never been
    what happens.

    Raising here propagates out of `lifespan`, so uvicorn exits non-zero and
    Kubernetes shows CrashLoopBackOff with the reason in `kubectl logs`. That
    is the loudest signal available, and a cluster that cannot authenticate
    anyone should be using it.
    """
    auth0 = get_settings().auth0
    errors = auth0.configuration_errors()

    if not auth0.required:
        if errors:
            logger.warning(
                "AUTH_REQUIRED=false: starting with an incomplete Auth0 configuration "
                "(%s). Every endpoint behind get_current_user/get_optional_user will "
                "return 503 (Authentication temporarily unavailable); no request will "
                "be authenticated in this deployment.",
                "; ".join(errors),
            )
        else:
            logger.warning(
                "AUTH_REQUIRED=false: Auth0 startup validation skipped, though the "
                "configuration looks complete."
            )
        return
    if errors:
        raise RuntimeError(
            "Auth0 configuration is incomplete or invalid; refusing to start. "
            + "; ".join(errors)
            + ". Fix this cluster's kustomize/config/<cluster>/api.env, or set "
            "AUTH_REQUIRED=false to run this deployment without authentication."
        )
    logger.info(
        "Auth0 configuration validated: issuer=%s audience=%s trusted_issuers=%d",
        auth0.issuer_url() if not auth0.has_explicit_trusted_issuers() else "(explicit map)",
        auth0.audience if not auth0.has_explicit_trusted_issuers() else "(per-issuer)",
        len(auth0.trusted_issuer_map()),
    )
    # AUTH-MAJ-004: the hosted password-change route converts ordinary token
    # possession into an account-changing capability, so a cluster that enables
    # it should also say, loudly, whether the recent-authentication gate is on.
    # Non-fatal by design -- the gate needs tenant work (a Post-Login Action that
    # stamps `auth_time`) that may not have landed -- but it must not be enabled
    # silently, which is exactly what an unremarked default would do.
    if auth0.password_reset_client_id and not auth0.password_reset_require_recent_auth:
        logger.warning(
            "Password reset is configured (AUTH0_PASSWORD_RESET_CLIENT_ID set) without recent-"
            "authentication enforcement (AUTH0_PASSWORD_RESET_REQUIRE_RECENT_AUTH is false): any "
            "valid eligible access token can mint a hosted password-change ticket. Enable the "
            "gate before exposing the reset UI; see docs/auth0-p2-decisions.md (D-12)."
        )

@asynccontextmanager
async def lifespan(_app: FastAPI) -> AsyncGenerator[None, None]:
    _validate_auth0_configuration()
    # #16: best-effort OIDC discovery warm. Deliberately after the (local,
    # side-effect-free) configuration gate and wrapped so it can never fail
    # startup -- a pod that cannot reach the discovery endpoint at boot still
    # starts and serves tokens via the convention-derived URLs.
    await warm_discovery_cache(get_settings().auth0)
    await init_standalone()
    yield
    await shutdown_standalone()


app = FastAPI(openapi_url="/openapi.json", docs_url=None, redoc_url=None, title=APP_TITLE, version=APP_VERSION, servers=APP_SERVERS, lifespan=lifespan)

# add origins
app.add_middleware(
    CORSMiddleware,
    allow_origins=APP_ORIGINS,
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"])

# include routers
app.include_router(compatibility_router)
app.include_router(simulations_router)
app.include_router(run_summary_router)
app.include_router(projects_router)
app.include_router(users_router)


def register_demo_router(app: FastAPI, *, enabled: bool) -> None:
    """Mount the /api/v1/demo/* RBAC teaching router only when explicitly enabled.

    #20: the demo router is a worked example, not production API surface. Left
    unmounted (the default), its paths return 404 and it is absent from
    /openapi.json. The Keycloak integration suite enables it via
    ENABLE_RBAC_DEMO in the test environment.
    """
    if enabled:
        app.include_router(rbac_demo_router)


register_demo_router(app, enabled=get_settings().enable_rbac_demo)


# -- endpoint logic -- #

@app.get("/")
def root() -> dict[str, str]:
    return {
        'docs': 'https://biosim.biosimulations.org/docs',
        'version': APP_VERSION
    }


@app.get("/version")
def get_version() -> str:
    return APP_VERSION


@app.get("/health", include_in_schema=False)
def health() -> dict[str, str]:
    # Liveness: deliberately dependency-free so a transient Mongo/Temporal
    # blip doesn't get the pod killed by a liveness probe.
    return {"status": "ok"}


@app.get("/ready", include_in_schema=False)
async def ready(
    response: Response,
    jwks_cache: JwksCache = Depends(get_jwks_cache),
) -> dict[str, object]:
    checks: dict[str, bool] = {}

    mongo_client = get_mongo_client()
    if mongo_client is None:
        checks["mongodb"] = False
    else:
        try:
            await mongo_client.admin.command("ping")
            checks["mongodb"] = True
        except Exception as e:
            logger.warning(f"/ready: mongodb ping failed: {e}")
            checks["mongodb"] = False

    checks["temporal"] = get_temporal_client() is not None

    ok = all(checks.values())
    response.status_code = 200 if ok else 503
    # #19c: auth (JWKS cache) health is reported as NON-GATING information --
    # it is deliberately kept out of `checks`, so `ok` stays computed from
    # MongoDB + Temporal only. A warm-cache Auth0 outage therefore never makes a
    # pod unready (it is still validating tokens), and this call makes no
    # outbound Auth0 request. Decision D-5: inform, do not gate.
    info: dict[str, object] = {"auth": jwks_cache.status()}
    return {"status": "ready" if ok else "not ready", "checks": checks, "info": info}


@app.get("/docs", include_in_schema=False)
async def custom_swagger_ui_html() -> HTMLResponse:
    if app.openapi_url is None:
        raise HTTPException(status_code=404, detail="OpenAPI schema not available")
    return get_swagger_ui_html(
        openapi_url=app.openapi_url,
        title=app.title + " - Swagger UI",
        swagger_js_url="https://cdn.jsdelivr.net/npm/swagger-ui-dist@5/swagger-ui-bundle.js",
        swagger_css_url="https://cdn.jsdelivr.net/npm/swagger-ui-dist@5/swagger-ui.css",
    )

@app.post(
    "/verify/omex",
    response_model=VerifyWorkflowOutput,
    operation_id="verify-omex",
    tags=["Verification"],
    dependencies=[Depends(get_temporal_client), Depends(get_file_service), Depends(get_local_cache_dir), Depends(get_omex_database_service), Depends(workflow_rate_limit)],
    summary="Request verification report for OMEX/COMBINE archive across simulators")
async def verify_omex(
        uploaded_file: UploadFile = File(..., description="OMEX/COMBINE archive containing a deterministic SBML model"),
        user: AuthenticatedUser = Depends(get_current_user),
        workflow_id_prefix: str = Query(default="omex-verification-", description="Prefix for the workflow id."),
        simulators: list[str] = Query(default=["amici", "copasi", "pysces", "tellurium", "vcell"],
                                      description="List of simulators 'name' or 'name:version' to compare."),
        include_outputs: bool = Query(default=False,
                                      description="Whether to include the output data on which the comparison is based."),
        user_description: str = Query(default="my-omex-compare", description="User description of the verification run."),
        rel_tol: float = Query(default=0.0001, description="Relative tolerance for proximity comparison."),
        abs_tol_min: float = Query(default=0.001, description="Min absolute tolerance, where atol = max(atol_min, max(arr1,arr2)*atol_scale."),
        abs_tol_scale: float = Query(default=0.00001, description="Scale for absolute tolerance, where atol = max(atol_min, max(arr1,arr2)*atol_scale."),
        cache_buster: str = Query(default="0", description="Optional unique id for cache busting (unique string to force new simulation runs)."),
        observables: Optional[list[str]] = Query(default=None,
                                                 description="List of observables to include in the return data.")
) -> VerifyWorkflowOutput:
    # ---- using hash to avoid saving multiple copies, upload to cloud storage if needed ---- #
    file_service = get_file_service()
    assert file_service is not None
    omex_database = get_omex_database_service()
    assert omex_database is not None
    omex_file: OmexFile = await get_cached_omex_file_from_upload(file_service=file_service, omex_database=omex_database,
                                                                 uploaded_file=uploaded_file, owner=user.sub)

    # ---- create workflow input ---- #
    simulator_versions: list[BiosimulatorVersion] = []
    biosim_service = get_biosim_service()
    assert biosim_service is not None
    all_simulator_versions = await biosim_service.get_simulator_versions()
    for simulator in simulators:
        simulator_version: Optional[BiosimulatorVersion] = None
        if ":" in simulator:
            name, version = simulator.split(":")
            for sv in all_simulator_versions:
                if sv.id == name and sv.version == version:
                    simulator_version = sv
                    break
        else:
            for sv in all_simulator_versions:
                if sv.id == simulator:
                    simulator_version = sv  # don't break, we want the last one in the list
        if simulator_version is not None:
            simulator_versions.append(simulator_version)
        else:
            raise HTTPException(status_code=400, detail=f"Simulator {simulator} not found.")

    workflow_id = f"{workflow_id_prefix}{uuid.uuid4()}"
    compare_settings = CompareSettings(user_description=user_description, include_outputs=include_outputs,
                                       rel_tol=rel_tol, abs_tol_min=abs_tol_min, abs_tol_scale=abs_tol_scale,
                                       observables=observables)
    omex_verify_workflow_input = OmexVerifyWorkflowInput(omex_file=omex_file, requested_simulators=simulator_versions,
                                                         compare_settings=compare_settings, cache_buster=cache_buster,
                                                         owner_sub=user.sub)

    # ---- invoke workflow ---- #
    # Log only non-sensitive correlation fields -- never the whole OmexFile
    # (its repr would carry the owner's raw Auth0 subject).
    logger.info(
        "starting verify workflow %s for OMEX hash %s (visibility=%s)",
        workflow_id,
        omex_file.file_hash_md5,
        omex_file.visibility,
    )
    temporal_client = get_temporal_client()
    assert temporal_client is not None
    workflow_handle = await temporal_client.start_workflow(
        OmexVerifyWorkflow.run,
        args=[omex_verify_workflow_input],
        task_queue="verification_tasks",
        id=workflow_id,
    )
    logger.info(f"started workflow with id {workflow_id}")
    assert workflow_handle.id == workflow_id

    # ---- return initial workflow output ---- #
    omex_verify_workflow_output = VerifyWorkflowOutput(
        compare_settings=compare_settings,
        workflow_status=VerifyWorkflowStatus.PENDING,
        timestamp=str(datetime.now(UTC)),
        workflow_id=workflow_id,
        workflow_run_id=workflow_handle.run_id,
        owner_sub=user.sub
    )
    return omex_verify_workflow_output


class _VerifyOwnership:
    """Adapter so ``require_owner_or_admin`` can check a verify workflow.

    Email is never persisted on the verify payload; ownership is ``owner_sub``
    only. Legacy in-flight workflows with ``owner_sub is None`` skip this check
    and allow any authenticated caller.
    """

    def __init__(self, owner_sub: str | None) -> None:
        self.owner_sub = owner_sub
        self.email: str | None = None


@app.get(
    "/verify/{workflow_id}",
    response_model=VerifyWorkflowOutput,
    operation_id='get-verify-output',
    name="Retrieve verification report",
    tags=["Verification"],
    dependencies=[Depends(get_temporal_client)],
    summary='Retrieve verification report for OMEX/COMBINE archive')
async def get_verify_output(
        workflow_id: str,
        user: AuthenticatedUser = Depends(get_current_user),
) -> VerifyWorkflowOutput:
    logger.info(f"in get /verify/{workflow_id}")

    try:
        # query temporal for the workflow output
        temporal_client = get_temporal_client()
        assert temporal_client is not None
        workflow_handle = temporal_client.get_workflow_handle(workflow_id=workflow_id,
                                                              result_type=VerifyWorkflowOutput)
        workflow_output: VerifyWorkflowOutput = await workflow_handle.query("get_output",
                                                                                result_type=VerifyWorkflowOutput,
                                                                                rpc_timeout=timedelta(seconds=60))
        if workflow_output.owner_sub is not None:
            require_owner_or_admin(
                user,
                [_VerifyOwnership(workflow_output.owner_sub)],
                action="read",
            )
        return workflow_output
    except HTTPException:
        raise
    except Exception as e2:
        exc_message = str(e2)
        msg = f"error retrieving verification job output with id: {workflow_id}: {exc_message}"
        logger.error(msg, exc_info=e2)
        raise HTTPException(status_code=404, detail=msg)


@app.post(
    "/verify/runs",
    response_model=VerifyWorkflowOutput,
    operation_id="verify-runs",
    tags=["Verification"],
    dependencies=[Depends(get_temporal_client), Depends(workflow_rate_limit)],
    summary="Request verification report for biosimulation runs by run IDs")
async def verify_runs(
        user: AuthenticatedUser = Depends(get_current_user),
        workflow_id_prefix: str = Query(default="runs-verification-", description="Prefix for the workflow id."),
        biosimulations_run_ids: list[str] = Query(default=["67817a2e1f52f47f628af971","67817a2eba5a3f02b9f2938d"],
                                                  description="List of biosimulations run IDs to compare."),
        include_outputs: bool = Query(default=False,
                                      description="Whether to include the output data on which the comparison is based."),
        user_description: str = Query(default="my-verify-job", description="User description of the verification run."),
        rel_tol: float = Query(default=0.0001, description="Relative tolerance for proximity comparison."),
        abs_tol_min: float = Query(default=0.001, description="Min absolute tolerance, where atol = max(atol_min, max(arr1,arr2)*atol_scale."),
        abs_tol_scale: float = Query(default=0.00001, description="Scale for absolute tolerance, where atol = max(atol_min, max(arr1,arr2)*atol_scale."),
        observables: Optional[list[str]] = Query(default=None,
                                                 description="List of observables to include in the return data.")
) -> VerifyWorkflowOutput:

    # ---- create workflow input ---- #
    workflow_id = f"{workflow_id_prefix}{uuid.uuid4()}"
    compare_settings = CompareSettings(user_description=user_description, include_outputs=include_outputs,
                                       rel_tol=rel_tol, abs_tol_min=abs_tol_min, abs_tol_scale=abs_tol_scale,
                                       observables=observables)
    runs_verify_workflow_input = RunsVerifyWorkflowInput(biosimulations_run_ids=biosimulations_run_ids,
                                                         compare_settings=compare_settings, owner_sub=user.sub)

    # ---- invoke workflow ---- #
    logger.info(f"starting verify workflow for biosim run IDs {biosimulations_run_ids}")
    temporal_client = get_temporal_client()
    assert temporal_client is not None
    workflow_handle = await temporal_client.start_workflow(
        RunsVerifyWorkflow.run,
        args=[runs_verify_workflow_input],
        task_queue="verification_tasks",
        id=workflow_id,
    )
    logger.info(f"started workflow with id {workflow_id}")
    assert workflow_handle.id == workflow_id

    # ---- return initial workflow output ---- #
    runs_verify_workflow_output = VerifyWorkflowOutput(
        compare_settings=compare_settings,
        workflow_status=VerifyWorkflowStatus.PENDING,
        timestamp=str(datetime.now(UTC)),
        workflow_id=workflow_id,
        workflow_run_id=workflow_handle.run_id,
        owner_sub=user.sub
    )
    return runs_verify_workflow_output


if __name__ == "__main__":
    uvicorn.run(app, host="0.0.0.0", port=8000)
    logger.info("Server started")