# backend/biosim_server/api/spec/openapi_3_1_0_generated.yaml
openapi: 3.1.0
info:
  title: biosim-server
  version: 0.10.0
paths:
...
  /verify/omex:
    post:
      tags:
      - Verification
      summary: Request verification report for OMEX/COMBINE archive across simulators
      operationId: verify-omex
      security:
      - HTTPBearer: []
      parameters:
      - name: workflow_id_prefix
        in: query
        required: false
        schema:
          type: string
          description: Prefix for the workflow id.
          default: omex-verification-
          title: Workflow Id Prefix
        description: Prefix for the workflow id.
      - name: simulators
        in: query
        required: false
        schema:
          type: array
          items:
            type: string
          description: List of simulators 'name' or 'name:version' to compare.
          default:
          - amici
          - copasi
          - pysces
          - tellurium
          - vcell
          title: Simulators
        description: List of simulators 'name' or 'name:version' to compare.
      - name: include_outputs
        in: query
        required: false
        schema:
          type: boolean
          description: Whether to include the output data on which the comparison
            is based.
          default: false
          title: Include Outputs
        description: Whether to include the output data on which the comparison is
          based.
      - name: user_description
        in: query
        required: false
        schema:
          type: string
          description: User description of the verification run.
          default: my-omex-compare
          title: User Description
        description: User description of the verification run.
      - name: rel_tol
        in: query
        required: false
        schema:
          type: number
          description: Relative tolerance for proximity comparison.
          default: 0.0001
          title: Rel Tol
        description: Relative tolerance for proximity comparison.
      - name: abs_tol_min
        in: query
        required: false
        schema:
          type: number
          description: Min absolute tolerance, where atol = max(atol_min, max(arr1,arr2)*atol_scale.
          default: 0.001
          title: Abs Tol Min
        description: Min absolute tolerance, where atol = max(atol_min, max(arr1,arr2)*atol_scale.
      - name: abs_tol_scale
        in: query
        required: false
        schema:
          type: number
          description: Scale for absolute tolerance, where atol = max(atol_min, max(arr1,arr2)*atol_scale.
          default: 1.0e-05
          title: Abs Tol Scale
        description: Scale for absolute tolerance, where atol = max(atol_min, max(arr1,arr2)*atol_scale.
      - name: cache_buster
        in: query
        required: false
        schema:
          type: string
          description: Optional unique id for cache busting (unique string to force
            new simulation runs).
          default: '0'
          title: Cache Buster
        description: Optional unique id for cache busting (unique string to force
          new simulation runs).
      - name: observables
        in: query
        required: false
        schema:
          anyOf:
          - type: array
            items:
              type: string
          - type: 'null'
          description: List of observables to include in the return data.
          title: Observables
        description: List of observables to include in the return data.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_verify-omex'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyWorkflowOutput'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /verify/{workflow_id}:
    get:
      tags:
      - Verification
      summary: Retrieve verification report for OMEX/COMBINE archive
      operationId: get-verify-output
      security:
      - HTTPBearer: []
      parameters:
      - name: workflow_id
        in: path
        required: true
        schema:
          type: string
          title: Workflow Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyWorkflowOutput'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /verify/runs:
    post:
      tags:
      - Verification
      summary: Request verification report for biosimulation runs by run IDs
      operationId: verify-runs
      security:
      - HTTPBearer: []
      parameters:
      - name: workflow_id_prefix
        in: query
        required: false
        schema:
          type: string
          description: Prefix for the workflow id.
          default: runs-verification-
          title: Workflow Id Prefix
        description: Prefix for the workflow id.
      - name: biosimulations_run_ids
        in: query
        required: false
        schema:
          type: array
          items:
            type: string
          description: List of biosimulations run IDs to compare.
          default:
          - 67817a2e1f52f47f628af971
          - 67817a2eba5a3f02b9f2938d
          title: Biosimulations Run Ids
        description: List of biosimulations run IDs to compare.
      - name: include_outputs
        in: query
        required: false
        schema:
          type: boolean
          description: Whether to include the output data on which the comparison
            is based.
          default: false
          title: Include Outputs
        description: Whether to include the output data on which the comparison is
          based.
      - name: user_description
        in: query
        required: false
        schema:
          type: string
          description: User description of the verification run.
          default: my-verify-job
          title: User Description
        description: User description of the verification run.
      - name: rel_tol
        in: query
        required: false
        schema:
          type: number
          description: Relative tolerance for proximity comparison.
          default: 0.0001
          title: Rel Tol
        description: Relative tolerance for proximity comparison.
      - name: abs_tol_min
        in: query
        required: false
        schema:
          type: number
          description: Min absolute tolerance, where atol = max(atol_min, max(arr1,arr2)*atol_scale.
          default: 0.001
          title: Abs Tol Min
        description: Min absolute tolerance, where atol = max(atol_min, max(arr1,arr2)*atol_scale.
      - name: abs_tol_scale
        in: query
        required: false
        schema:
          type: number
          description: Scale for absolute tolerance, where atol = max(atol_min, max(arr1,arr2)*atol_scale.
          default: 1.0e-05
          title: Abs Tol Scale
        description: Scale for absolute tolerance, where atol = max(atol_min, max(arr1,arr2)*atol_scale.
      - name: observables
        in: query
        required: false
        schema:
          anyOf:
          - type: array
            items:
              type: string
          - type: 'null'
          description: List of observables to include in the return data.
          title: Observables
        description: List of observables to include in the return data.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyWorkflowOutput'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
...
    Body_verify-omex:
      properties:
        uploaded_file:
          type: string
          format: binary
          title: Uploaded File
          description: OMEX/COMBINE archive containing a deterministic SBML model
      type: object
      required:
      - uploaded_file
      title: Body_verify-omex
    CompareSettings:
      properties:
        user_description:
          type: string
          title: User Description
        include_outputs:
          type: boolean
          title: Include Outputs
        rel_tol:
          type: number
          title: Rel Tol
        abs_tol_min:
          type: number
          title: Abs Tol Min
        abs_tol_scale:
          type: number
          title: Abs Tol Scale
        observables:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Observables
      type: object
      required:
      - user_description
      - include_outputs
      - rel_tol
      - abs_tol_min
      - abs_tol_scale
      title: CompareSettings
    ComparisonStatistics:
      properties:
        dataset_name:
          type: string
          title: Dataset Name
        simulator_version_i:
          type: string
          title: Simulator Version I
        simulator_version_j:
          type: string
          title: Simulator Version J
        var_names:
          items:
            type: string
          type: array
          title: Var Names
        score:
          anyOf:
          - items:
              type: number
            type: array
          - type: 'null'
          title: Score
        is_close:
          anyOf:
          - items:
              type: boolean
            type: array
          - type: 'null'
          title: Is Close
        error_message:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Message
      type: object
      required:
      - dataset_name
      - simulator_version_i
      - simulator_version_j
      - var_names
      title: ComparisonStatistics
...
    GenerateStatisticsActivityOutput:
      properties:
        sims_run_info:
          items:
            $ref: '#/components/schemas/SimulationRunInfo'
          type: array
          title: Sims Run Info
        comparison_statistics:
          additionalProperties:
            items:
              items:
                $ref: '#/components/schemas/ComparisonStatistics'
              type: array
            type: array
          type: object
          title: Comparison Statistics
        sim_run_data:
          anyOf:
          - items:
              $ref: '#/components/schemas/RunData'
            type: array
          - type: 'null'
          title: Sim Run Data
      type: object
      required:
      - sims_run_info
      - comparison_statistics
      title: GenerateStatisticsActivityOutput
...
    HDF5File:
      properties:
        filename:
          type: string
          title: Filename
        id:
          type: string
          title: Id
        uri:
          type: string
          title: Uri
        groups:
          items:
            $ref: '#/components/schemas/HDF5Group'
          type: array
          title: Groups
      type: object
      required:
      - filename
      - id
      - uri
      - groups
      title: HDF5File
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    RunData:
      properties:
        run_id:
          type: string
          title: Run Id
        dataset_name:
          type: string
          title: Dataset Name
        var_names:
          items:
            type: string
          type: array
          title: Var Names
        data:
          $ref: '#/components/schemas/Hdf5DataValues'
      type: object
      required:
      - run_id
      - dataset_name
      - var_names
      - data
      title: RunData
...
    SimulationRunInfo:
      properties:
        biosim_sim_run:
          $ref: '#/components/schemas/BiosimSimulationRun'
        hdf5_file:
          $ref: '#/components/schemas/HDF5File'
      type: object
      required:
      - biosim_sim_run
      - hdf5_file
      title: SimulationRunInfo
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    VerifyWorkflowOutput:
      properties:
        workflow_id:
          type: string
          title: Workflow Id
        compare_settings:
          $ref: '#/components/schemas/CompareSettings'
        workflow_status:
          $ref: '#/components/schemas/VerifyWorkflowStatus'
        timestamp:
          type: string
          title: Timestamp
        workflow_run_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Workflow Run Id
        workflow_error:
          anyOf:
          - type: string
          - type: 'null'
          title: Workflow Error
        workflow_results:
          anyOf:
          - $ref: '#/components/schemas/GenerateStatisticsActivityOutput'
          - type: 'null'
        owner_sub:
          anyOf:
          - type: string
          - type: 'null'
          title: Owner Sub
      type: object
      required:
      - workflow_id
      - compare_settings
      - workflow_status
      - timestamp
      title: VerifyWorkflowOutput
    VerifyWorkflowStatus:
      type: string
      enum:
      - PENDING
      - IN_PROGRESS
      - COMPLETED
      - FAILED
      - RUN_ID_NOT_FOUND
      title: VerifyWorkflowStatus
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer